floating everything

+++ src/milesToMuppets/data.py
@@ -1,5 +1,5 @@
 # default_mph_speed = 670_616_629
-default_mph_speed = 60
+default_mph_speed = 60.0
 data = {
         "songs": {
             "Muppets Most Wanted": "3Z1dw9cLeFAyhvkXdn6P5G",
@@ -11,7 +11,7 @@ data = {
             1: "The Muppets"
         },
         "constants": {
-            "defMphSpeed": default_mph_speed,
-            "defMinPerMile": 60 / default_mph_speed
+            "defMphSpeed": float(default_mph_speed),
+            "defMinPerMile": float(60 / default_mph_speed)
         }
     }
